Refactoring to reduce cognitive complexity
