+++ src/projectSetup/projectSetup.jsx
@@ -1,12 +1,22 @@
 const DEBUG_MODE_ON = false
 
-const shell = require('shelljs')
 const React = require('react')
 
-const {useState, useEffect} = React
+const {useEffect} = React
 const {string, shape, oneOf, func, bool, arrayOf} = require('prop-types')
 const {Box, Text, Color} = require('ink')
-const handlebars = require('handlebars')
+const {
+  useStep,
+  useStartSetup,
+  useCloneRepoStep,
+  useStartGitProject,
+  useCreateLicense,
+  useCreateCodeOfConduct,
+  useCreateReadme,
+  useUpdatePackageJson,
+  useInstallPackages,
+  useCreateGitCommit
+} = require('./hooks')
 const importJsx = require('import-jsx')
 
 const Task = importJsx('../components/task.jsx')
@@ -15,333 +25,78 @@ const {licenses, codesOfConduct} = require('../enum')
 const SetupProject = ({
   configuration,
   onHasFinished
-// TODO: Refactor this
-// eslint-disable-next-line sonarjs/cognitive-complexity
 }) => {
-  const [
+  const [step, onNextStep] = useStep()
+
+  useStartSetup({
+    onNextStep,
+    hasConfigurationArrived: (
+      Object.keys(configuration).length > 0
+    )
+  })
+
+  const isGitClonePending = useCloneRepoStep({
     step,
-    onSetStep
-  ] = useState(0)
+    onNextStep,
+    packageName: configuration.packageName
+  })
 
-  const onNextStep = () => onSetStep(step + 1)
+  const isGitProjectPending = useStartGitProject({
+    step,
+    onNextStep,
+    packageName: configuration.packageName,
+    gitRepoUrl: configuration.gitRepoUrl
+  })
 
-  useEffect(() => {
-    if (Object.keys(configuration).length > 0) {
-      onNextStep()
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [configuration])
+  const isLicensePending = useCreateLicense({
+    step,
+    onNextStep,
+    license: configuration.license,
+    authorName: configuration.authorName
+  })
 
-  // Step 1:
-  // Clone boilerplate git repo
+  const isCodeOfConductPending = useCreateCodeOfConduct({
+    step,
+    onNextStep,
+    codeOfConduct: configuration.codeOfConduct,
+    authorEmail: configuration.authorEmail
+  })
 
-  const [
-    isGitClonePending,
-    onSetIsGitClonePending
-  ] = useState(false)
+  const isReadmePending = useCreateReadme({
+    step,
+    onNextStep,
+    packageName: configuration.packageName,
+    description: configuration.description,
+    isScoped: configuration.isScoped,
+    scopeName: configuration.scopeName,
+    codeOfConduct: configuration.codeOfConduct,
+    license: configuration.license
+  })
 
-  useEffect(() => {
-    if (step === 1) {
-      // Clone GitHub repo with boilerplate
-      onSetIsGitClonePending(true)
-      shell.exec([
-        'git clone',
-        'https://github.com/bence-toth/react-library-boilerplate.git',
-        configuration.packageName
-      ].join(' '), {
-        async: true,
-        silent: true
-      }, () => {
-        // Finished git clone
-        onSetIsGitClonePending(false)
-        onNextStep()
-      })
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
+  const isPackageJsonPending = useUpdatePackageJson({
+    step,
+    onNextStep,
+    isScoped: configuration.isScoped,
+    scopeName: configuration.scopeName,
+    packageName: configuration.packageName,
+    description: configuration.description,
+    keywords: configuration.keywords,
+    license: configuration.license,
+    authorName: configuration.authorName,
+    authorEmail: configuration.authorEmail,
+    authorWebsite: configuration.authorWebsite,
+    gitRepoUrl: configuration.gitRepoUrl
+  })
 
-  // Step 2:
-  // Delete existing `.git` folder
-  // Start new git project
-  // Optional: Add git remote
+  const isNpmInstallPending = useInstallPackages({
+    step,
+    onNextStep
+  })
 
-  const [
-    isGitProjectPending,
-    onSetIsGitProjectPending
-  ] = useState(false)
-
-  useEffect(() => {
-    if (step === 2) {
-      onSetIsGitProjectPending(true)
-      shell.cd(configuration.packageName)
-      // Delete old git folder
-      shell.rm('-rf', '.git')
-      // Start new git project with git init
-      shell.exec('git init', {
-        async: true,
-        silent: true
-      }, () => {
-        // Finished git init
-        if (configuration.gitRepoUrl.length > 0) {
-          // Add git remote if URL was provided
-          shell.exec(`git remote add origin ${configuration.gitRepoUrl}.git`)
-        }
-        onSetIsGitProjectPending(false)
-        onNextStep()
-      })
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
-
-  // Step 3:
-  // Optional: Add license
-  const [
-    isLicensePending,
-    onSetIsLicensePending
-  ] = useState(false)
-
-  useEffect(() => {
-    if (step === 3) {
-      onSetIsLicensePending(true)
-      // Delete old LICENSE file
-      shell.rm('-rf', 'LICENSE')
-      if (configuration.license === licenses.mit) {
-        // Create MIT license file
-        const templateCode = shell
-          .cat(`${__dirname}/templates/licenses/mit.hbr`)
-          .toString()
-        const template = handlebars.compile(templateCode)
-        const licenseContent = template({
-          year: `${new Date().getFullYear()}`,
-          copyrightOwner: configuration.authorName
-        })
-        const licensePath = './LICENSE'
-        shell.ShellString(licenseContent).to(licensePath)
-        onSetIsLicensePending(false)
-        onNextStep()
-      }
-      else {
-        onSetIsLicensePending(false)
-        onNextStep()
-      }
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
-
-  // Step 4:
-  // Optional: Add Code of Conduct
-  const [
-    isCodeOfConductPending,
-    onSetIsCodeOfConductPending
-  ] = useState(false)
-
-  useEffect(() => {
-    if (step === 4) {
-      onSetIsCodeOfConductPending(true)
-      if (configuration.codeOfConduct === codesOfConduct.contributorCovenant) {
-        // Create CoC file
-        const templateCode = shell
-          .cat(`${__dirname}/templates/codesOfConduct/contributorCovenant.hbr`)
-          .toString()
-        const template = handlebars.compile(templateCode)
-        const codeOfConductContent = template({
-          email: configuration.authorEmail
-        })
-        const codeOfConductPath = './code-of-conduct.md'
-        shell.ShellString(codeOfConductContent).to(codeOfConductPath)
-        onSetIsCodeOfConductPending(false)
-        onNextStep()
-      }
-      else {
-        onSetIsCodeOfConductPending(false)
-        onNextStep()
-      }
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
-
-  // Step 5:
-  // Add readme file
-  const [
-    isReadmePending,
-    onSetIsReadmePending
-  ] = useState(false)
-
-  useEffect(() => {
-    if (step === 5) {
-      onSetIsReadmePending(true)
-      const templateCode = shell
-        .cat(`${__dirname}/templates/readme.hbr`)
-        .toString()
-      const availableScripts = shell
-        .cat('./README.md')
-        .toString()
-      const template = handlebars.compile(templateCode, {noEscape: true})
-      const readmeContent = template({
-        packageName: configuration.packageName,
-        hasDescription: configuration.description.length > 0,
-        description: configuration.description,
-        fullPackageName: (
-          configuration.isScoped
-            ? `${configuration.scopeName}/${configuration.packageName}`
-            : configuration.packageName
-        ),
-        hasCodeOfConduct: configuration.codeOfConduct !== null,
-        availableScripts,
-        hasLicense: configuration.license !== null,
-        license: configuration.license
-      })
-      const readmePath = './README.md'
-      shell.ShellString(readmeContent).to(readmePath)
-      onSetIsReadmePending(false)
-      onNextStep()
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
-
-  // Step 6:
-  // Update `package.json`
-  const [
-    isPackageJsonPending,
-    onSetIsPackageJsonPending
-  ] = useState(false)
-
-  useEffect(() => {
-    if (step === 6) {
-      onSetIsPackageJsonPending(true)
-      const packageJsonPath = './package.json'
-      const {
-        scripts,
-        files,
-        main,
-        devDependencies,
-        peerDependencies,
-        dependencies
-      } = JSON.parse((
-        shell
-          .cat(packageJsonPath)
-          .toString()
-      ))
-      const updatedPackageObject = {
-        name: (
-          configuration.isScoped
-            ? `${configuration.scopeName}/${configuration.packageName}`
-            : configuration.packageName
-        ),
-        version: '1.0.0',
-        ...(
-          (configuration.description.length > 0)
-            ? {
-              description: configuration.description
-            }
-            : {}
-        ),
-        ...(
-          (configuration.keywords.length > 0)
-            ? {
-              keywords: configuration.keywords
-            }
-            : {}
-        ),
-        ...(
-          (configuration.license === licenses.mit)
-            ? {
-              license: 'MIT'
-            }
-            : {}
-        ),
-        scripts,
-        author: {
-          name: configuration.authorName,
-          ...(
-            (configuration.authorEmail.length > 0)
-              ? {
-                email: configuration.authorEmail
-              }
-              : {}
-          ),
-          ...(
-            (configuration.authorWebsite.length > 0)
-              ? {
-                url: configuration.authorWebsite
-              }
-              : {}
-          )
-        },
-        ...(
-          (configuration.gitRepoUrl.length > 0)
-            ? {
-              homepage: `${configuration.gitRepoUrl}#readme`,
-              repository: {
-                type: 'git',
-                url: `${configuration.gitRepoUrl}.git`
-              },
-              bugs: {
-                url: `${configuration.gitRepoUrl}/issues`
-              }
-            }
-            : {}
-        ),
-        dependencies,
-        peerDependencies,
-        devDependencies,
-        main,
-        files
-      }
-      shell
-        .ShellString(JSON.stringify(updatedPackageObject, null, 2))
-        .to(packageJsonPath)
-      onSetIsPackageJsonPending(false)
-      onNextStep()
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
-
-  // Step 7:
-  // Install packages
-  const [
-    isNpmInstallPending,
-    onSetIsNpmInstallPending
-  ] = useState(false)
-
-  useEffect(() => {
-    if (step === 7) {
-      onSetIsNpmInstallPending(true)
-      shell.exec('npm install', {
-        async: true,
-        silent: true
-      }, () => {
-        // Finished npm install
-        onSetIsNpmInstallPending(false)
-        onNextStep()
-      })
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
-
-
-  // Step 8:
-  // Create git commit
-  const [
-    isGitCommitPending,
-    onSetIsGitCommitPending
-  ] = useState(false)
-
-  useEffect(() => {
-    if (step === 8) {
-      onSetIsGitCommitPending(true)
-      shell.exec('git add -A', {silent: true})
-      shell.exec('git commit -m "Initial commit"', {
-        async: true,
-        silent: true
-      }, () => {
-        // Finished npm install
-        onSetIsGitCommitPending(false)
-        onNextStep()
-      })
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [step])
+  const isGitCommitPending = useCreateGitCommit({
+    step,
+    onNextStep
+  })
 
   useEffect(() => {
     if (step === 9) {
